requirements.txt
asyncio
python-binance
